Consolidate and move loader tests to test/unit/test_loader.py
